+++ pytorch_lightning/utilities/cli.py
@@ -37,7 +37,8 @@ if _JSONARGPARSE_AVAILABLE:
 
     set_config_read_mode(fsspec_enabled=True)
 else:
-    ArgumentParser = Namespace = object
+    locals()["ArgumentParser"] = object
+    locals()["Namespace"] = object
 
 
 class _Registry(dict):